+++ app/client/lib/menu_item.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class MenuItem {
+  String name;
+  IconData icon;
+  MaterialColor color;
+
+  MenuItem(this.name, this.icon, this.color);
+}
